chore: move utilities into separate file
